packages/apps/Music: Fix for the clip skip at the end of playback
-While playing a clip, during the last 2-3 seconds playabck if the
user selects the same clip from the playlist, next clip is getting
selected by skipping the playback of last 2-3 seconds of the current
clip.

-In Music App, when play is called, a check is done on the
current clip duration. If the current clip has 2 seconds of playback
time remaining, the next clip is getting selected for playback.
Because of this the last 2-3 seconds of playback is getting skipped
for the current clip.

-Fixing this issue by reducing the check on remaining time duration
to 1 sec for the curernt clip.

CRs-Fixed: 332750
(cherry picked from commit 42437c48842a461a269e79e6a0a37373aafa06d7)

Change-Id: I9c8a3c82c02d13ce3e1865ca263b1966104c7e49