Music: Fix for seekbar issue on interrupt by call
- Seek bar stucks on call interrupt while fast-forwarding
or rewinding a clip.
- When user presses next/previous for seeking, scanForward()
is called by fwdListener which in turn performs seek operation
updating timeline and seekbar. Each event of seek is posted to
the listener with a delay of 260ms.
In this usecase when the target gets the call, by the time seek event
is posted with repcnt as -1, mService is set to NULL.Since mService
is null it returns without setting mPosoverride to -1.Due to this,
mPosoverride retains the older value where the last seek operation
was successfully completed.When call is disconnected and music
playback is resumed, since mPosoverride is positive, position
of the seekbar is always updated with mPosoverride in refreshNow().
- In scanForward() and scanBackward(), set mPosoverride to -1
in case mService is null and is called with repcnt -1.
Since seek event is posted with a delay, this ensures that
last seek event is acknowledged even if mService is null.

Change-Id: I527be16537ee95b754581047ab7cb1e0c09de7a8

CRs-Fixed: 349576