Music: Fix for seek bar stucks after pressing suspend key
- Seek bar stucks while pressing Suspend key during Fast-forwarding
or Rewinding the clip
- During power suspend, mPosoverride value is set to the current
playback position since seeking is in progress. mPosoverride is
reset only when user stops pressing seek button. In this use case
Suspend is called before getting the last seek event. Hence seek bar
is always at a constant position taking the value of mPosoverride.
- Reset mPosoverride to -1 when suspend key is pressed if
mPosoverride is set to a positive value.

Change-Id: I6f572a93db2236971ce65e01d69be5bb2c7e7b45

CRs-Fixed: 354251